--- execute/exit_status.h
@@ -1,10 +1,0 @@
-#ifndef EXIT_STATUS_H
-#define EXIT_STATUS_H
-
-#define ES_IS_A_DIRECTORY 126
-#define ES_PERMISSION_DENIED 126
-#define ES_COMMAND_NOT_FOUND 127
-#define ES_NO_SUCH_FILE 127
-# define ES_SYNTAX_ERROR 258
-
-#endif //EXIT_STATUS_H
